pyproject.toml
[tool.poetry]
name = "django-pastebin"
version = "0.1.0"
description = ""
authors = ["BlaackWizard <kabynks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.68"
django-storages = "^1.14.4"
psycopg2-binary = "^2.9.10"
django = "^5.1.3"
django-ninja = "^1.3.0"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
